% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl}
\alias{cl}
\title{cl - linear model clustering function}
\usage{
cl(dat, fm, cluster)
}
\arguments{
\item{dat}{data frame from which model was fitted}

\item{fm}{fitted model including varible to be clustered}

\item{cluster}{string, variable by which to cluster standard errors}
}
\description{
cl - linear model clustering function
}
