% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveRunner.R
\name{curveRunner}
\alias{curveRunner}
\title{curveRunner - processes s-curve objects}
\usage{
curveRunner(s.curve.model, inc.tidy.models = NULL,
  inc.full.models = NULL, perm.index = NULL)
}
\arguments{
\item{s.curve.model}{Specification curve model (generated by s.curve function)}

\item{inc.tidy.models}{logical, should results include tidy model output for each specification (from package broom)? If specified, will override default passed from s.curve.model object (e.g., for simpler output in permutation testing)}

\item{inc.full.models}{logical, should results include default model output for each specification (e.g., from function lm)? If specified, will override default passed from s.curve.model object (e.g., for simpler output in permutation testing)}

\item{perm.index}{index value of permutation testing.  If present, index column will be added to results.}
}
\value{
data frame, with rows containing model estimation results for each individual model specification in overall s-curve model
}
\description{
curveRunner - processes s-curve objects
}
