% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.curve.plot.R
\name{s.curve.plot}
\alias{s.curve.plot}
\title{Plotting tool for specification curve output from s.curve}
\usage{
s.curve.plot(s.curve.mod, what = "estimate", title = NULL,
  plot.order = NULL, decreasing = TRUE, show.sig = TRUE,
  include.permutations = FALSE, perm.sig = FALSE, pointsize = 2,
  x.alt = NULL, plot.theme = NULL)
}
\arguments{
\item{s.curve.mod}{fitted model object from function s.curve() or
s.curve.update()}

\item{what}{character, for each specification, what result is to be plotted?
Accepts "estimate" (treatment estimate, default), "p.value", "perm.p.value"
(permutation-test-estimated p-value), "AIC", "BIC"}

\item{title}{character, what should the title of the plot be?}

\item{plot.order}{character - what, if anything, should the values be sorted
on? Accepts same values as estimate "parameter"}

\item{decreasing}{logical, when sorting, use decreasing order (default TRUE)}

\item{show.sig}{logical, should the points be colorized by significance?
(default TRUE)}

\item{include.permutations}{logical, if permutations are present, include CI
boundaries of permutations? (default FALSE)}

\item{perm.sig}{logical, use permutation p-values for significance markings?
(default FALSE)}

\item{pointsize}{numeric, size of points to use in plotting (default 2)}

\item{x.alt}{numeric vector, with same length as # specifications.  If
included, manually positions specifications}

\item{plot.theme}{character, name of theme available to ggplot to override
default (e.g. "minimal", "bw", "classic")}
}
\value{
a ggplot plot summarizing the results of the fitted s.curve object
  passed to it.  Can be saved, or will render plot in viewer if printed to
  console.
}
\description{
Plotting tool for specification curve output from s.curve
}
