% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.curve.R
\name{s.curve}
\alias{s.curve}
\title{Specification curve maker function}
\usage{
s.curve(dat, outcomes, treatment, cov.list, no.cov.exclude = NULL,
  extra.models = NULL, extra.treatment = NULL, mod.type = lm,
  mod.family = NULL, alpha = 0.05, tail = NULL, subsets = NULL,
  subsets.exclude = TRUE, weights = NULL, weights.exclude = TRUE,
  weights.ipw.vars = NULL, permutations = NULL, perm.pvalues = FALSE,
  perm.seed = NULL, cluster = FALSE, cluster.var = NULL,
  robust.se = NULL, cat.percent = TRUE, keep.tidy.models = TRUE,
  keep.full.models = FALSE, model.only = FALSE)
}
\arguments{
\item{dat}{a dataframe containing the variables}

\item{outcomes}{vector of outcome variables}

\item{treatment}{vector of treatment variables}

\item{cov.list}{named list, sets of covariates / moderators - write
moderators as "var1:var2"; eg: cov.list = list(gender = c("gender.roster",
"gender.selfreport")}

\item{no.cov.exclude}{Will add a model where each item of the list above is
missing, unless specified here}

\item{extra.models}{Models written literally; will be appended to the set
(still crossed against subsets and weights)}

\item{extra.treatment}{Treatment variable in extra models (length 1 or same
as other, needed if doing inverse probability weighting}

\item{mod.type}{takes lm, glm}

\item{mod.family}{if family needed}

\item{alpha}{nominal alpha value (if one tailed, will test against p <= .1
on that side)}

\item{tail}{One tailed test?  Takes "upper" and "lower"}

\item{subsets}{runs subsets of data based on the specifications listed here
(vector of conditions)}

\item{subsets.exclude}{if subsets added, includes an un-subsetted version}

\item{weights}{vector of weight variables names to add to runs, re-named if
desired or, a weighting method (currenly accepts "ipw.calc" for inverse
probability weighting)}

\item{weights.exclude}{Includes an unweighted version where weights added}

\item{weights.ipw.vars}{in addition to treatment variable, any other (factor
or categorical) variables to consider when (re)weighting subsets}

\item{permutations}{optional, number of permutations for p-curve}

\item{perm.pvalues}{logical, calculates permutation test-based pvalues (if
permutation test active)}

\item{perm.seed}{optional, RNG seed to use for permutation test (integer,
will create one if not present)}

\item{cluster}{optional, Cluster robust standard errors}

\item{cluster.var}{optional, Variable on which to cluster, string}

\item{robust.se}{optional, heteroskedasticity-consistent standard error
adjustment provided by package "sandwich".  See details in ?vcovHC.}

\item{cat.percent}{logical, displays summary output of % significant at the
end of the data for convenience in interactive mode. Set to false if using
as a part of an Rmarkdown file.}

\item{keep.tidy.models}{logical, set to false for large model samples}

\item{keep.full.models}{logical, set to false for large model samples}

\item{model.only}{logical, outputs the model for later use, rather than
running (default is FALSE)}
}
\value{
depending on paramters, either a fitted s-curve object or a template for fitting an s-curve object
}
\description{
Specification curve maker function
}
