% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.curve.density.plot.R
\name{s.curve.sig.density.plot}
\alias{s.curve.sig.density.plot}
\title{Significance density plot for permutation testing:}
\usage{
s.curve.sig.density.plot(s.curve.mod, subtitle.add = NULL)
}
\arguments{
\item{s.curve.mod}{fitted s-curve model}

\item{subtitle.add}{subtitle to add to main title}
}
\value{
Outputs a ggplot object that plots a density curve of significance on the
}
\description{
Significance density plot for permutation testing:
}
