% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.curve.subset.R
\name{s.curve.subset}
\alias{s.curve.subset}
\title{Utility function for subsetting specification curves based on specification characteristics}
\usage{
s.curve.subset(s.curve.mod, ...)
}
\arguments{
\item{s.curve.mod}{fitted specification curve model object}

\item{...}{character, logical statements that can be used to filter the specification based on the results}
}
\value{
calls s.curve.update and returns results and output from that function, including permutation tests if appropriate
}
\description{
Utility function for subsetting specification curves based on specification characteristics
}
