% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.curve.table.R
\name{s.curve.table}
\alias{s.curve.table}
\title{Creates a summary table of specification curve results (for export to spreadsheet)}
\usage{
s.curve.table(s.curve.mod, ordering = NULL)
}
\arguments{
\item{s.curve.mod}{fitted specification curve model}

\item{ordering}{character, choice of orderings from the model.  Accepts
"estimate", "p.value", "perm.p.value", "AIC", "BIC"}
}
\value{
dataframe containing the treatment parameter estimates, statistics,
  and permutation test results for each specification
}
\description{
Creates a summary table of specification curve results (for export to spreadsheet)
}
