% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.curve.update.R
\name{s.curve.update}
\alias{s.curve.update}
\title{Calculates output & incorporates overall results, including permutation test data if present}
\usage{
s.curve.update(s.curve.mod, perm.dat = NULL, perm.pvalues = NULL)
}
\arguments{
\item{s.curve.mod}{estimated s.curve model from curveRunner}

\item{perm.dat}{permutation test data to add (if run externally)}

\item{perm.pvalues}{logical, calculate p-values against permutation test data? (if NULL, uses default in model object)}
}
\value{
updated s-curve model with results summaries (also printed to console)
}
\description{
Calculates output & incorporates overall results, including permutation test data if present
}
